[MOD]  networkType:  [ADD] network
